Increase camera resolution, improve window scaling, round ball coordinates, and adjust ball radius for better visibility

+++ src/Video.cpp
@@ -10,7 +10,7 @@ cv::Mat addBallsToPicture(cv::Mat frame) {
     for (Ball& ball : ballList) {
         ball.move(1.0f, cols); // Bewege den Ball, jetzt mit Breite
         if (ball.x >= 0 && ball.x < cols && ball.y >= 0 && ball.y < rows) {
-            cv::circle(frame, cv::Point(ball.getX(), ball.getY()), ball.radius, ball.getColor(), -1); // Ball zeichnen
+            cv::circle(frame, cv::Point(static_cast<int>(ball.getX()), static_cast<int>(ball.getY())), ball.radius, ball.getColor(), -1); // Ball zeichnen
         }
     }
     return frame;
